Re #26375 Update SANS tests to use new var names
