+++ Testing/Data/UnitTest/USER_ZOOM_MONITOR_4_AND_5_SHIFTED.txt.md5
@@ -1,0 +1,1 @@
+186a52643c13344d2af4b7909a92b576
