+++ scripts/SANS/sans/gui_logic/models/state_gui_model.py
@@ -513,9 +513,6 @@ class StateGuiModel(ModelCommon):
     def z_offset(self, value):
         self.set_simple_element(element_id=SampleId.offset, value=value)
 
-
-
-
     # ==================================================================================================================
     # ==================================================================================================================
     # Q TAB
